Drop Aeson support, drop external pieces.
Types no longer derive aeson implementations. The user (a library
including this one) is repsonsible for taking care of that.

As a side effect, TranspilablePack no longer accept a list of external
pieces. It was never used anyway, but forced users to derive aeson
instances for PieceOfLogic, which complicates things a bit.

External pieces may return in some other form, if deemed necessary.